+++ src-tauri/build.rs
@@ -2,6 +2,7 @@ fn main() {
     tauri_build::try_build(tauri_build::Attributes::new().app_manifest(
         tauri_build::AppManifest::new().commands(&[
             "watch_state",
+            "update_action",
             "start_diagnostic",
             "cancel_operation",
             "open_external",